Learn HTML by Building a Cat Photo App - Step 8
HTML attributes are special words used inside the opening tag of an element to control the element's behavior. 
The src attribute in an img element specifies the image's URL (where the image is located). 
An example of an img element using an src attribute: <img src="https://www.example.com/the-image.jpg">.

Inside the existing img element, add an src attribute with this URL:

https://cdn.freecodecamp.org/curriculum/cat-photo-app/relaxing-cat.jpg