Learn Basic CSS by Building a Cafe Menu - Step 58
It is a bit boring for all the text to have the same font-family.
You can still have the majority of the text sans-serif and make just
the h1 and h2 elements different using a different selector.

Style both the h1 and the h2 elements so that only these elements'
text use Impact font.