Learn Basic CSS by Building a Cafe Menu - Step 79
The default color of a link that has not yet been clicked on is typically blue.
The default color of a link that has already been visited from a page is typically purple.

To make the footer links the same color regardless if a link has been visited,
use a type selector for the anchor element (a) and use the value black for the color property.