Learn Basic CSS by Building a Cafe Menu - Step 84
The menu text CAMPER CAFE has a different space from the top than the
address's space at the bottom of the menu. This is due to the browser
having some default top margin for the h1 element.

Change the top margin of the h1 element to 0 to remove all the top margin.