Learn Basic CSS by Building a Cafe Menu - Step 89
The image you added is not centered horizontally like the Coffee heading
above it. img elements are "like" inline elements.

To make the image behave like heading elements (which are block-level),
create an img type selector and use the value block for the display
property and use the applicable margin-left and margin-right values
to center it horizontally.