Learn CSS Colors Build Colored Marker - Step 03
Remember that the title element gives search engines extra information
about the page. It also tells browsers what text to display in the title
bar when the page is open, and on the tab for the page.

Within the head element, nest a title element with the text
Colored Markers.