Learn CSS Colors Build Colored Marker - Step 16
To give the markers different colors, you will need to add a unique
class to each one. Multiple classes can be added to an element
by listing them in the class attribute and separating them with a space.
For example, the following adds both the animal and dog classes to a
div element.

<div class="animal dog">
To begin, add the class one to the first marker div element.