Learn CSS Colors Build Colored Marker - Step 31
Now that you're familiar with secondary colors,
you'll learn how to create tertiary colors. Tertiary colors are created
by combining a primary with a nearby secondary color.

To create the tertiary color orange, update the rgb function
in the .one CSS rule so that red is at the max value, and set green to 127.