Learn CSS Colors Build Colored Marker - Step 32
Notice that, to create orange, you had to increase the intensity of red
and decrease the intensity of the green rgb values. This is because orange
is the combination of red and yellow, and falls between the two colors on the color wheel.

To create the tertiary color spring green, combine cyan with green.
Update the rgb function in the .two CSS rule so that green is at the max value, and set blue to 127.