Learn CSS Colors Build Colored Marker - Step 34
There are three more tertiary colors: chartreuse green (green + yellow),
 azure (blue + cyan), and rose (red + magenta).

To create chartreuse green, update the rgb function in the .one rule
so that red is at 127, and set green to the max value.

For azure, update the rgb function in the .two rule so that green is at 127
and blue is at the max value.

And for rose, which is sometimes called bright pink, update the rgb
function in the .three rule so that blue is at 127 and red is at the max value.