Learn CSS Colors Build Colored Marker - Step 35
Now that you've gone through all the primary, secondary, and tertiary
colors on a color wheel, it'll be easier to understand other
color theory concepts and how they impact design.

First, in the rules .one, .two, and .three, adjust the values in the
rgb function so that the background-color of each element is set to pure black.
Remember that the rgb function uses the additive color model,
where colors start as black and change as the values of red, green, and blue increase.