Learn CSS Colors Build Colored Marker - Step 57
Now that the color-stops are set, you'll apply different shades of red
to each color argument in the linear-gradient function. The shades
on the top and bottom edges of the marker will be darker,
while the one in the middle will be lighter, as if there's a light above it.

For the first color argument, which is currently pure red, update the
rgb function so the value for red is 122, the value for green is 74, and the value for blue is 14.