Learn CSS Colors Build Colored Marker - Step 74
Another way to set the opacity for an element is with the alpha channel.

Similar to the opacity property, the alpha channel controls how transparent or opaque a color is.

You've already set sleeve's opacity with a named color and the opacity property,

but you can add an alpha channel to the other CSS color properties.

Inside the .sleeve rule, remove the opacity property and value.