Learn CSS Colors Build Colored Marker - Step 75
You're already familiar with using the rgb function to set colors.

To add an alpha channel to an rgb color, use the rgba function instead.

The rgba function works just like the rgb function, but takes one more number from 0 to 1.0 for the alpha channel:

rgba(redValue, greenValue, blueValue, alphaValue);

In the .sleeve rule, use the rgba function to set the background-color property to pure white with 50% opacity.