Learn CSS Colors Build Colored Marker - Step 76
Your sleeve is looking good, but it would look even better if it was positioned

more toward the right side of the marker. One way to do that is to add

another element before the sleeve to push it to the right.

Add a new div with the class cap before the sleeve div element.