Learn CSS Colors Build Colored Marker - Step 92
The shadows for your green and blue markers will have the same position, blur, and spread.

The only difference will be the colors.

In the .green and .blue CSS rules, update the values for the box-shadow properties so offsetX is 0,offsetY is 0, blurRadius is 20px, and spreadRadius is 0. Leave the colors as green and blue for now.