Learn HTML Forms Build Registration Form - Step 33
Adding a dropdown to the form is easy with the select element.

The select element is a container for a group of option elements, and the option element acts as a label for each dropdown option. Both elements require closing tags.

Start, by adding a select element below the two label elements. Then, nest 5 option elements within the select element.