Learn CSS Flexbox Building Photo Gallery - Step 21
The alt image attribute should describe the image content. Screen readers announce the alternative text in place of images. If the image can't be loaded, this text is presented in place of the image.

To complete the project, add an alt attribute to all nine of your cat images to describe them. Use a value at least five characters long for each.
	modified:   flexBoxPhotoGallery/index.html