Build Balance Sheet PseudoSelector - Step 21
Within the first tr, add a th element with the text Loans The outstanding balance on our startup loan.. Wrap that text, except for Loans , within a span element with the class set to description.

Add three td elements below that, and give them the following text, in order: $500, $250, and $0. Give the third td element a class set to current.