Build Balance Sheet PseudoSelector - Step 22
Within the second tr, add a th element with the text Expenses Annual anticipated expenses, such as payroll.. Wrap that text, except for Expenses , within a span element with the class set to description.

Add three td elements below that, and give them the following text, in order: $200, $300, and $400. Give the third td element a class set to current.