Build Balance Sheet PseudoSelector - Step 23
Within the third tr, add a th element with the text Credit The outstanding balance on our credit card.. Wrap that text, except for Credit , within a span element with the class set to description.

Add three td elements below that, and give them the following text, in order: $50, $50, and $75. Give the third td element a class set to current.