Build Picasso Painting Intermed. CSS - Step 10
The z-index property is used to create "layers" for your HTML elements. If you are familiar with image editing tools, you may have worked with layers before. This is a similar concept.

Elements with a higher z-index value will appear to be layered on top of elements with a lower z-index value. This can be combined with the positioning in the previous lesson to create unique effects.

Since the back-wall element will need to appear "behind" the other elements you will be creating, give the back-wall element a z-index of -1.