Build Picasso Painting Intermed. CSS - Step 60
Within the #guitar element, create three div elements. Give the first two a class value of guitar. Then give the first an id of guitar-left, and the second an id of guitar-right. Add an id to the third div with the value guitar-neck.

The third div should not have the guitar class.